Day 1 scan iterator (#2)
* Migrate day1 to lib to add tests

**This Commit**
Moves the functional code for `day` to a `lib.rs`.

**Why?**
So we can add tests and refactor fearlessly!

* Add benchmarks

**This Commit**
Adds a `criterion` benchmark suite for `day1`.

**Why?**
So we can see how various refactorings affect performance.

* Use `std::iter::Iterator::scan`

**This Commit**
Uses the built in [`scan`][scan] method on iterators instead of using
the custom iterator.

**Why?**
I originally thought it'd be cleaner and shorter. I forgot that `scan`
needed to return an `Option` so now there's a weird return at the bottom
of `process_floor` so I'm not so sure anymore. You be the judge!

[scan]: https://doc.rust-lang.org/std/iter/trait.Iterator.html#method.scan