Remove unused panic case
**This Commit**
Removes the `panic!` fallback case when processing floor instructions.

**Why?**
Mostly to test performance. In production code this wouldn't be the best
idea but since the problem gives us a well formed input we can assume
that and allow the compiler to be awesome.

**Results**
```
part1                   time:   [1.3374 us 1.3447 us 1.3551 us]
                        change: [-92.914% -92.840% -92.771%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low mild
  6 (6.00%) high mild
  5 (5.00%) high severe

part2                   time:   [1.0825 us 1.0928 us 1.1063 us]
                        change: [-59.226% -58.769% -58.253%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe
```

**Assembly Details**
What I _thought_ would happen was that it'd have a branchless
conditional move. What it _actually_ did was compare the input byte to
`40` (which is `b'('`). It then sets the `al` register to 1 if they are
equal and `0` if they are not. `al` is the [last byte of the `eax`
register][register]. It then doubles the value in `eax` (which, from
the last step is either 0 or 1 because of how we set `al` to 0 or 1)
so it is now either 0 or 2. Then it subtracts 1 to get to 1 or -1.

[godbolt link](https://godbolt.org/z/G3orq4sMc)
[register]: https://www.cs.virginia.edu/~evans/cs216/guides/x86.html