Document non-improvement of using `sum` for day1
**This Commit**
Adds a quick note to `part1` documenting a failed attempt at further
optimization.

**Why?**
Instead of using `.scan().last()` for `part1` I wondered if it'd be more
efficient to sum up all the numbers and then do all the math:

```rust
let total = input.bytes().map(|b| b as i32).sum::<i32>();
-(total * 2 - 81 * input.len() as i32)
```

but this turned out to be the exact same runtime. I left myself a note
because I could easily see myself forgetting I tried this and
implementing it again but I didn't want to make the code change because
it does nothing and it increases the duplicated code by not leveraging
`process_floor`.