Actually use `$path` in `aoc_input!`
**This Commit**
Updates the `aoc_input` macro to use the passed `$path` when determining
which file to read.

**Why?**
Because the original was definitely a typo. However, we could decide
that all inputs are in a consistent place and not have the macro take an
argument but this seems less weird for now.

See https://github.com/mlodato517/aoc_2015/pull/6#discussion_r810288814
for more details.